services/search-service/requirements.txt
fastapi==0.110.2
uvicorn[standard]==0.29.0
strawberry-graphql==0.211.0
pydantic==2.6.4
pydantic-settings==2.2.1
elasticsearch==8.12.1
httpx==0.27.0
python-dateutil==2.9.0.post0
